feat: Remove lint module and add spotify search and display result
This commit removes the custom lint module and implements spotify search and display result.

- Removed `lint` module.
- Removed lint dependencies in `build.gradle.kts`.
- Removed `lint` in `settings.gradle.kts`.
- Removed `lint.api`, `lint.checks`, `lint.tests`, `lint` in `libs.versions.toml`
- Remove `TestMethodNameDetector`, `DesignSystemDetector`, `HRKIssueRegistry` and all test for lint
- Update `build.gradle.kts` for each module.
- Add `SearchSpotifyBody`, `TracksItem` for spotify response.
- Update `ApiService` add search track.
- Update `KeyRequest`:
  - add `SEARCH_SPOTIFY`
- Update `SearchingRepository` and `SearchingRepositoryImpl`:
  - change `youtubeSearching` to `search`
- Update `TuneDetectViewModel`:
  - add init for searchingRepository
  - add `token` for searching in spotify
  - add function `searchResultFromSpotify`
  - add `spotifyResponse`
- Update `bottom_sheet`:
  - add `SpotifyResultComponent`, `SpotifyItemComponent` to display result from spotify.
  - add height for view
  - add `convertMillisecondsToMinutes`
- Add code style config.
- Update `libs.versions.toml` to update lib and plugin.
- Update `DynamicAsyncImage` to update `contentScale` and `imageLoader` parameter.
- Update `ACRCloudResponse` to update `Empty_Result` parameter.
- Update `MainActivity`: add log for response from acrcloud.

+++ core/model/src/main/kotlin/com/hrk/apps/hrkdev/core/model/spotify/SearchSpotifyResponse.kt
@@ -19,7 +19,7 @@ data class TracksItem(
     val artists: List<ArtistX>? = null,
     val available_markets: List<String>? = null,
     val disc_number: Int? = null,
-    val duration_ms: Int? = null,
+    val duration_ms: Long? = null,
     val explicit: Boolean? = null,
     val external_ids: ExternalIds? = null,
     val external_urls: ExternalUrlsXXX? = null,
